Throw 501 not implemented on unrecognized method
