+++ WS++/HTTP/Method.h
@@ -5,7 +5,7 @@
 #include <iostream>
 
 #include "../Common.h"
-#include "HTTPErrors.h"
+#include "Errors.h"
 
 namespace HTTPMethod
 {
