+++ WS++/main.cpp
@@ -3,7 +3,7 @@
 #include "Common.h"
 #include "Sockets/AsyncSocketServer.h"
 
-#include "HTTP/HTTPRequest.h"
+#include "HTTP/Request.h"
 
 void handle_socket(Socket& socket)
 {
@@ -13,6 +13,10 @@ void handle_socket(Socket& socket)
 
 		HTTP::Request req(reader);
 		std::cout << "A request to " << req.path << std::endl;
+	}
+	catch (HTTP::ErrorCode & err)
+	{
+		std::cout << "HTTP/1.1 " << err.status() << " " << err.to_string() << std::endl;
 	}
 	catch (HTTP::Error & err)
 	{
